+++ src/main/java/com/datumbox/framework/machinelearning/common/bases/baseobjects/BaseTrainable.java
@@ -18,13 +18,15 @@ package com.datumbox.framework.machinelearning.common.bases.baseobjects;
 import com.datumbox.common.dataobjects.Dataframe;
 import com.datumbox.common.objecttypes.Trainable;
 import com.datumbox.common.persistentstorage.interfaces.DatabaseConfiguration;
+import com.datumbox.framework.machinelearning.common.dataobjects.StandardKnowledgeBase;
 import com.datumbox.framework.machinelearning.common.dataobjects.KnowledgeBase;
+import java.io.Serializable;
 import java.lang.reflect.InvocationTargetException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * Base class for every Model of the Framework. This includes Machine Learning
+ * Base class for every Trainable Algorithm of the Framework. This includes Machine Learning
  * Models, Data Transformers, Feature Selectors etc.
  *
  * @author Vasilis Vryniotis <bbriniotis@datumbox.com>
@@ -32,7 +34,7 @@ import org.slf4j.LoggerFactory;
  * @param <TP>
  * @param <KB>
  */
-public abstract class BaseTrainable<MP extends BaseModelParameters, TP extends BaseTrainingParameters, KB extends KnowledgeBase<MP, TP>> implements Trainable<MP, TP> {
+public abstract class BaseTrainable<MP extends BaseModelParameters, TP extends BaseTrainingParameters, KB extends StandardKnowledgeBase<MP, TP>> implements Trainable<MP, TP> {
     
     /**
      * The Logger of all algorithms.
@@ -43,13 +45,12 @@ public abstract class BaseTrainable<MP extends BaseModelParameters, TP extends B
     /**
      * The name of the Database where we persist our data.
      */
-    protected String dbName;
+    protected final String dbName;
     
     /**
-     * The KnowledgeBase instance of the algorithm. Do NOT access it directly,
-     * use the kb() getter instead.
+     * The KnowledgeBase instance of the algorithm. 
      */
-    protected KB knowledgeBase;
+    private final KB knowledgeBase;
     
     /**
      * Generates a new instance of a BaseTrainable by providing the Class of the
@@ -58,13 +59,13 @@ public abstract class BaseTrainable<MP extends BaseModelParameters, TP extends B
      * @param <BT>
      * @param aClass
      * @param dbName
-     * @param dbConfig
+     * @param dbConf
      * @return 
      */
-    public static <BT extends BaseTrainable> BT newInstance(Class<BT> aClass, String dbName, DatabaseConfiguration dbConfig) {
+    public static <BT extends BaseTrainable> BT newInstance(Class<BT> aClass, String dbName, DatabaseConfiguration dbConf) {
         BT algorithm = null;
         try {
-            algorithm = aClass.getConstructor(String.class, DatabaseConfiguration.class).newInstance(dbName, dbConfig);
+            algorithm = aClass.getConstructor(String.class, DatabaseConfiguration.class).newInstance(dbName, dbConf);
         } 
         catch (InstantiationException | IllegalAccessException | IllegalArgumentException | InvocationTargetException | NoSuchMethodException | SecurityException ex) {
             throw new RuntimeException(ex);
@@ -74,34 +75,23 @@ public abstract class BaseTrainable<MP extends BaseModelParameters, TP extends B
     }
 
     /**
-     * Protected Constructor which does not include the initialization of the
-     * KnowledgeBase.
+     * The basic Constructor of all BaseTrainable classes.
      * 
-     * @param dbName
+     * @param baseDBname
      * @param dbConf 
+     * @param kbClass 
+     * @param kbSubtypeClasses 
      */
-    protected BaseTrainable(String dbName, DatabaseConfiguration dbConf) {
+    protected BaseTrainable(String baseDBname, DatabaseConfiguration dbConf, Class<? extends StandardKnowledgeBase> kbClass, Class<? extends Serializable>... kbSubtypeClasses) {
         String methodName = this.getClass().getSimpleName();
         String dbNameSeparator = dbConf.getDBnameSeparator();
-        if(!dbName.contains(methodName+dbNameSeparator)) { //patch for the K-fold cross validation which already contains the name of the algorithm in the dbname
-            dbName += dbNameSeparator + methodName;
+        if(!baseDBname.contains(methodName+dbNameSeparator)) { //patch for the K-fold cross validation which already contains the name of the algorithm in the dbname
+            baseDBname += dbNameSeparator + methodName;
         }
         
-        this.dbName = dbName;
-    }
-    
-    /**
-     * Protected Constructor which includes the initialization of the
-     * KnowledgeBase.
-     * 
-     * @param dbName
-     * @param dbConf
-     * @param mpClass
-     * @param tpClass 
-     */
-    protected BaseTrainable(String dbName, DatabaseConfiguration dbConf, Class<MP> mpClass, Class<TP> tpClass) {
-        this(dbName, dbConf);
-        this.knowledgeBase = (KB) new KnowledgeBase(this.dbName, dbConf, mpClass, tpClass);
+        dbName = baseDBname;
+        
+        knowledgeBase = (KB) KnowledgeBase.newInstance(kbClass, dbName, dbConf, kbSubtypeClasses);
     }
     
     /** {@inheritDoc} */