+++ src/main/java/com/datumbox/framework/machinelearning/common/bases/datatransformation/DataTransformer.java
@@ -22,7 +22,7 @@ import com.datumbox.common.persistentstorage.interfaces.DatabaseConnector;
 
 import com.datumbox.framework.machinelearning.common.bases.baseobjects.BaseModelParameters;
 import com.datumbox.framework.machinelearning.common.bases.baseobjects.BaseTrainingParameters;
-import com.datumbox.framework.machinelearning.common.dataobjects.KnowledgeBase;
+import com.datumbox.framework.machinelearning.common.dataobjects.StandardKnowledgeBase;
 
 /**
  * Base class for all the Data Transformers of the framework.
@@ -31,7 +31,7 @@ import com.datumbox.framework.machinelearning.common.dataobjects.KnowledgeBase;
  * @param <MP>
  * @param <TP>
  */
-public abstract class DataTransformer<MP extends DataTransformer.ModelParameters, TP extends DataTransformer.TrainingParameters> extends BaseTrainable<MP, TP, KnowledgeBase<MP, TP>> {
+public abstract class DataTransformer<MP extends DataTransformer.ModelParameters, TP extends DataTransformer.TrainingParameters> extends BaseTrainable<MP, TP, StandardKnowledgeBase<MP, TP>> {
 
     /**
      * The ModelParameters class stores the coefficients that were learned during
@@ -65,7 +65,7 @@ public abstract class DataTransformer<MP extends DataTransformer.ModelParameters
      * @see com.datumbox.framework.machinelearning.common.bases.baseobjects.BaseTrainable#BaseTrainable(java.lang.String, com.datumbox.common.persistentstorage.interfaces.DatabaseConfiguration, java.lang.Class, java.lang.Class)  
      */
     protected DataTransformer(String dbName, DatabaseConfiguration dbConf, Class<MP> mpClass, Class<TP> tpClass) {
-        super(dbName, dbConf, mpClass, tpClass);
+        super(dbName, dbConf, StandardKnowledgeBase.class, mpClass, tpClass);
     }
     
     /**